change namespace to the controller which admin is going to create
