+++ app/views/admin/users/new.html.erb
@@ -1,4 +1,4 @@
-<%= form_with(model: @added_user, url: added_users_path) do |form| %>
+<%= form_with(model: @admin_user, url: admin_users_path) do |form| %>
   <div class="container">
     <div class="row justify-content-center">
       <div class="col-md-8">
