+++ app/views/admin/users/show.html.erb
@@ -12,7 +12,7 @@
   </div>
 
   <div class="mt-3">
-    <%= link_to 'Edit', edit_added_user_path(@user), class: "btn btn-primary" %>
-    <%= link_to 'Back', added_users_path, class: "btn btn-secondary" %>
+    <%= link_to 'Edit', edit_admin_user_path(@user), class: "btn btn-primary" %>
+    <%= link_to 'Back', admin_users_path, class: "btn btn-secondary" %>
   </div>
 </div>
